put py files in separated cosmology/modeling/dataops directories
